Fix resizing losing scroll bar
This would happen when opening menu, resizing, closing, then reopening.
The scroll bar would be gone. My best guess for why occurred is a timing
issue from when it ran in `populateEmotesMenu` to the time it ran during
the reopening of the menu. Also moved adding the classes during resize
after the resizing is finished; it will reduce some lag during resize,
as well as make it more apparent at the end of resizing that it becomes
pinned (resize then tack it there).