Add Twitch Turbo emotes
This is a half-hard-coded implementation. Emotes will always appear if
the user has the set. Known turbo sets will be hard-coded to have the
channel as "turbo". Even if the turbo sets were not hard-coded, the
emotes would appear, but with no grouping.

Wrapped in a try...catch just in case Ember isn't being nice.

A known effect, although low priority and probably impossible to fix, is
turbo emotes won't load until after chat has loaded.