Ensure menu UI created on initial load (#4)
* Ensure menu UI created on initial load

This addresses a case where plugins try to load elements before the UI
dom has been created by ensuring the DOM is created up front and
attached/reattached later as needed.
https://github.com/night/BetterTTV/issues/2901

* Ensure updateEmotes() on attach()

The previous changes to replace usage of init() with new attach() method
did not include the call to updateEmotes() within attach(). This call
was made in init() and not including it can cause an issue where the
Favorites group does not show up. This change fixes that potential
issue.

* Show favorites group

Previous commit, while useful, didn't actually resolve it. Testing this
thing is so troublesome, I guess maybe I was seeing the live version
thinking it was the test. Because the UIMenu dom isn't attached when the
UIFavoritesGroup() constructor is called it can't find the
#starred-emotes-group. This changes it to look for that under the
UIMenu's dom, which will find it even while unattached.

* Fix to make resize handle work with changes

The resize handle wasn't working with the changes to how loading vs
attaching is working. Jquery UI was searching within the document for
the handle. This change passes the resize handle using a direct
reference to finding it within the UI menu dom.